requirements.txt
Itsdangerous==1.1.0
flask==1.1.1
jinja2==3.0.3
werkzeug==0.16.1

Flask-Cors==3.0.10
Flask-RESTful==0.3.7
SQLAlchemy==1.4.31
Flask-SQLAlchemy==2.5.1
APScheduler==3.8.1
Flask-APScheduler==1.12.3
alembic~=1.6.5
Flask-Migrate==3.1.0
gevent-websocket==0.10.1
concurrent-log-handler==0.9.16
requests==2.23.0
pandas==1.2.0
openpyxl==3.0.5
jieba==0.42.1
python-dateutil==2.8.2
qrcode==7.3.1
PyJWT==1.7.1
pycryptodome==3.9.0
PyYAML==6.0

Redis==3.3.1
pymongo==3.10.1
mysql-connector-python==8.0.18

pillow==9.0.0
#docker 环境不安装
# opencv-python==4.5.1.48


gunicorn==20.1.0

# 开发工具
pylint==2.6.0
# pipdeptree
